+++ sample/iosApp/Gradient/SharedView.swift
@@ -1,5 +1,4 @@
 import Composables
-import Models
 import SwiftUI
 
 struct SharedView: View {